+++ src/pages/mypage/MyHome.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+import MyPageContent from "../../components/content/MyPageContent";
+
+const MyHome = () => {
+
+    return (
+        <>
+            <MyPageContent>
+                <h3>홈</h3>
+            </MyPageContent>
+        </>
+    );
+};
+
+export default MyHome;
