+++ src/pages/mypage/MyLikes.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+import MyPageContent from "../../components/content/MyPageContent";
+
+const MyLikes = () => {
+
+    return (
+        <>
+            <MyPageContent>
+                <h3>좋아요</h3>
+            </MyPageContent>
+        </>
+    );
+};
+
+export default MyLikes;
