+++ src/pages/mypage/MySchedules.js
@@ -1,0 +1,18 @@
+import React from "react";
+
+import MyPageContent from "../../components/content/MyPageContent";
+
+const MySchedules = () => {
+
+
+    return (
+        <>
+            <MyPageContent>
+                <h2>나의 일정</h2>
+            </MyPageContent>
+        </>
+    );
+};
+
+
+export default MySchedules;
