+++ src/pages/mypage/MyTravels.js
@@ -1,0 +1,18 @@
+import React from "react";
+
+import MyPageContent from "../../components/content/MyPageContent";
+
+const MyTravels = () => {
+
+
+    return (
+        <>
+            <MyPageContent>
+                <h2>나의 여행</h2>
+            </MyPageContent>
+        </>
+    );
+};
+
+
+export default MyTravels;
