+++ src/pages/settings/ReportSettings.js
@@ -1,0 +1,17 @@
+import React from "react";
+import SettingContent from "../../components/content/SettingContent";
+
+const ReportSettings = () => {
+
+
+    return (
+        <>
+            <SettingContent>
+                신고
+            </SettingContent>
+        </>
+    );
+};
+
+
+export default ReportSettings;
